+++ test/run/library.js
@@ -29,7 +29,7 @@ export default function () {
     const { mode, codes, dependencies } = await falco(config)
 
     assert(mode === 'production')
-    assert(codes.map(s => s.name).join() === 'index.js,index.js.map')
+    assert(codes.map((s) => s.name).join() === 'index.js,index.js.map')
     assert(dependencies.join() === 'react')
   })
 })